+++ src/components/global/data/Banner.js
@@ -1,5 +1,7 @@
 import React from 'react';
 
+// Utilisé dans service & portfolio
+
 const Banner = () => {
   return <img src="/assets/images/banner.jpg" alt="banner" className="img-fluid" />;
 };
