+++ src/components/global/utils/UnderlineHover.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 
-//Appliquer un effect gras au survol
-//Utilisé sur le footer.
+// Appliquer un effect gras au survol
+// Utilisé sur le footer.
 
 const UnderlineHover = ({ className, children }) => {
   const [hovered, setHovered] = useState(false);
