+++ src/components/global/utils/isLinkElement.js
