+++ src/components/homePage/data/Bar.js
