+++ src/components/homePage/style/home.css
@@ -1,7 +1,7 @@
 /* MainHome */
 
 .backgroundImage {
-    background-image:url(../../../public/assets/images/hero-bg.jpg);
+    background-image:url(../../../../public/assets/images/hero-bg.jpg);
     background-repeat: no-repeat;
     background-size: cover;
     background-position: center;
