+++ src/pages/Contact.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Nav from '../components/nav/Nav';
-import Footer from '../components/footer/Footer';
 import HomeContact from '../components/contact/HomeContact';
+import Footer from '../components/footer/Footer';
 
 const Contact = () => {
   return (
